requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apptools==5.2.1
av==11.0.0
Cartopy==0.22.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
envisage==7.0.3
eva-decord==0.6.1
exceptiongroup==1.2.0
fastapi==0.109.0
fonttools==4.47.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mayavi==4.8.1
networkx==3.2.1
numpy==1.26.3
packaging==23.2
parameterized==0.9.0
pillow==10.2.0
portalocker==2.8.2
pydantic==2.6.0
pydantic_core==2.16.1
pyface==8.0.0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.6.7
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.1
traits==6.4.3
traitsui==8.0.0
types-regex==2023.12.25.20240106
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
vtk==9.3.0
wcwidth==0.2.13
yacs==0.1.8
zipp==3.17.0
annotated-types==0.6.0
anyio==4.2.0
apptools==5.2.1
av==11.0.0
Cartopy==0.22.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
envisage==7.0.3
eva-decord==0.6.1
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.0
fonttools==4.47.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mayavi==4.8.1
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.1
parameterized==0.9.0
pillow==10.2.0
portalocker==2.8.2
pydantic==2.6.0
pydantic_core==2.16.1
pyface==8.0.0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.6.7
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.1
traits==6.4.3
traitsui==8.0.0
types-regex==2023.12.25.20240106
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.0.post1
vtk==9.3.0
wcwidth==0.2.13
yacs==0.1.8
zipp==3.17.0
annotated-types==0.6.0
anyio==4.2.0
apptools==5.2.1
av==11.0.0
Cartopy==0.22.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
envisage==7.0.3
eva-decord==0.6.1
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.0
fonttools==4.47.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mayavi==4.8.1
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.1
parameterized==0.9.0
pillow==10.2.0
portalocker==2.8.2
pydantic==2.6.0
pydantic_core==2.16.1
pyface==8.0.0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.6.7
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.1
traits==6.4.3
traitsui==8.0.0
types-regex==2023.12.25.20240106
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.0.post1
vtk==9.3.0
wcwidth==0.2.13
yacs==0.1.8
zipp==3.17.0
